+++ tgbotapi.extensions.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/extensions/UpdateChatRetriever.kt
@@ -1,5 +1,7 @@
 package dev.inmo.tgbotapi.extensions.utils.extensions
 
+import dev.inmo.tgbotapi.CommonAbstracts.FromUser
+import dev.inmo.tgbotapi.CommonAbstracts.WithUser
 import dev.inmo.tgbotapi.extensions.utils.asFromUser
 import dev.inmo.tgbotapi.extensions.utils.asUser
 import dev.inmo.tgbotapi.extensions.utils.shortcuts.chat
@@ -12,20 +14,24 @@ import dev.inmo.tgbotapi.types.update.abstracts.Update
 import dev.inmo.tgbotapi.utils.PreviewFeature
 
 @PreviewFeature
-fun Update.sourceChat(): Chat? = when (this) {
-    is MediaGroupUpdate -> when (this) {
+fun Update.sourceChat(): Chat? = when {
+    this is MediaGroupUpdate -> when (this) {
         is SentMediaGroupUpdate -> data.chat
         is EditMediaGroupUpdate -> data.chat
     }
-    is BaseMessageUpdate -> data.chat
-    is InlineQueryUpdate -> data.from
-    is ChosenInlineResultUpdate -> data.user
-    is CallbackQueryUpdate -> data.user
-    is PreCheckoutQueryUpdate -> data.user
-    is PollAnswerUpdate -> data.user
-    is ShippingQueryUpdate -> data.user
-    else -> null
+    this is BaseMessageUpdate -> data.chat
+    else -> {
+        when (val data = data) {
+            is FromUser -> data.from
+            is WithUser -> data.user
+            else -> null
+        }
+    }
 }
 
 @PreviewFeature
-fun Update.sourceUser(): User? = data.asFromUser()?.user ?: sourceChat()?.asUser()
+fun Update.sourceUser(): User? = when (val data = data) {
+    is FromUser -> data.from
+    is WithUser -> data.user
+    else -> sourceChat()?.asUser()
+}